Proper support for arrays/objects
Previously, complex (array or object) types were "flattened" (e.g. 4
separate params for log topics instead of a single array param).

This was confusing from the UI perspective, but also made it harder to
maintain the flattened versions of these params.

Here, we implement a more robust and flexible solution, from the data
schema typing changes, to components, validation, and formatting.

The change will make it easier to add methods with complex params in
the future.